feat: 사용중인 카테고리인지 검사하는 로직 추가

+++ src/main/java/com/poortorich/category/service/CategoryService.java
@@ -1,5 +1,6 @@
 package com.poortorich.category.service;
 
+import com.poortorich.accountbook.service.AccountBookService;
 import com.poortorich.category.domain.model.enums.DefaultExpenseCategory;
 import com.poortorich.category.domain.model.enums.DefaultIncomeCategory;
 import com.poortorich.category.entity.Category;
@@ -29,6 +30,7 @@ public class CategoryService {
 
     private final CategoryRepository categoryRepository;
     private final UserService userService;
+    private final AccountBookService accountBookService;
 
     public List<DefaultCategoryResponse> getDefaultCategories(CategoryType type, String username) {
         return categoryRepository.findByUserAndType(userService.findUserByUsername(username), type).stream()
@@ -141,7 +143,13 @@ public class CategoryService {
     }
 
     public Response deleteCategory(Long id, String username) {
-        Category category = getCategoryOrThrow(id, userService.findUserByUsername(username));
+        User user = userService.findUserByUsername(username);
+        Category category = getCategoryOrThrow(id, user);
+
+        if (accountBookService.isUsedCategory(user, category, category.getType())) {
+            return CategoryResponse.NOT_DELETE_USED_CATEGORY;
+        }
+
         categoryRepository.delete(category);
 
         return CategoryResponse.DELETE_CATEGORY_SUCCESS;
